delete a course

+++ models/Course.js
@@ -1,4 +1,4 @@
-// const HttpError = require('http-errors');
+const HttpError = require('http-errors');
 const log = require('loglevel');
 const { db } = require('../services/database');
 // eslint-disable-next-line no-unused-vars -- TEMP FOR ESLINT
@@ -9,31 +9,17 @@ const { whereParams, insertValues } = require('../services/sqltools');
 // if successful, but not deleted, throw error
 // if db error, db.query will throw a rejected promise
 // otherwise throw error
-async function remove(criteria) {
-  const { text, params } = whereParams(criteria);
-  const res = await db.query(`DELETE FROM "course" ${text} RETURNING *;`, params);
-  if (res.rows.length > 0) {
-    log.debug(
-      `Successfully deleted course from db with criteria: ${text}, ${JSON.stringify(params)}`
-    );
-    return { success: true };
+async function remove(Id) {
+  if (Id) {
+    const { text, params } = whereParams({ id: Id });
+    const res = await db.query(`DELETE FROM "course" ${text} RETURNING *;`, params);
+    if (res.rows.length > 0) {
+      return (`Successfully deleted course from db`);
+    }
+    throw HttpError(500,'Unexpected db condition, delete successful with no returned record');
+  } else {
+    throw HttpError(400, 'Id is required.');
   }
-  log.debug(`No courses found in db with criteria: ${text}, ${JSON.stringify(params)}`);
-  return {};
-}
-
-// if found return [ {}, {} ... ]
-// if not found return []
-// if db error, db.query will throw a rejected promise
-async function findAll(criteria, limit = 100, offset = 0) {
-  const { text, params } = whereParams(criteria);
-  const n = params.length;
-  const p = params.concat([limit, offset]);
-  const res = await db.query(`SELECT * from "course" ${text} LIMIT $${n + 1} OFFSET $${n + 2};`, p);
-  log.debug(
-    `Retrieved ${res.rows.length} courses from db with criteria: ${text}, ${JSON.stringify(params)}`
-  );
-  return res.rows;
 }
 
 // if found return { ... }
@@ -50,36 +36,8 @@ async function findOne(criteria) {
   return {};
 }
 
-// // if successful insert return inserted record {}
-// // if successful, but no row inserted, throw error
-// // if db error, db.query will throw a rejected promise
-// // otherwise throw error
-// async function create(userId, email) {
-//   // userId and email are required
-//   if (userId && email) {
-//     const enable = email === env.masterAdminEmail;
-//     const role = email === env.masterAdminEmail ? 'admin' : 'user';
-//     const { text, params } = insertValues({
-//       userId: userId,
-//       email: email,
-//       enable: enable,
-//       role: role,
-//     });
-//     const res = await db.query(`INSERT INTO "user" ${text} RETURNING *;`, params);
-//     if (res.rows.length > 0) {
-//       log.debug(
-//         `Successfully inserted user ${email} into db with data: ${text}, ${JSON.stringify(params)}`
-//       );
-//       return res.rows[0];
-//     }
-//     throw HttpError(500, 'Unexpected DB Condition, insert sucessful with no returned record');
-//   } else {
-//     throw HttpError(400, 'UserId and Email are required.');
-//   }
-// }
 
 module.exports = {
   remove,
-  findOne,
-  findAll
+  findOne
 };
